LayoutTests/http/tests/security/contentSecurityPolicy/1.1/reportonly-in-meta-ignored.html

<!DOCTYPE html>
<html>
<head>
    <meta http-equiv="Content-Security-Policy-Report-Only" content="this-should-be-ignored">
    <script>
        if (window.testRunner)
            testRunner.dumpAsText();
    </script>
</head>
<body>
    <p>This tests that report-only Content Security Policy headers are ignored if contained in meta elements.</p>
</body>
</html>
