LayoutTests/http/tests/security/contentSecurityPolicy/1.1/reportonly-in-meta-ignored-expected.txt

CONSOLE MESSAGE: line 4: The report-only Content Security Policy 'this-should-be-ignored' was delivered via a <meta> element, which is disallowed. The policy has been ignored.
This tests that report-only Content Security Policy headers are ignored if contained in meta elements.
